+++ src/collective/js/jqueryui/example/jqueryui.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-from Products.Five import BrowserView
-
-
-class Example(BrowserView):
-    """Simple Example View with jqueryui example"""